Add functionality for box constraints (and remove it from formhandler)
