+++ cashocs/_forms/control_form_handler.py
@@ -64,10 +64,6 @@ class ControlFormHandler(form_handler.FormHandler):
         self.controls = optimization_problem.controls
         self.riesz_scalar_products = optimization_problem.riesz_scalar_products
         self.control_bcs_list = optimization_problem.control_bcs_list
-        self.control_constraints = optimization_problem.control_constraints
-        self.require_control_constraints = (
-            optimization_problem.require_control_constraints
-        )
 
         self.control_dim = len(self.controls)
         self.control_spaces: List[fenics.FunctionSpace] = [
@@ -228,151 +224,6 @@ class ControlFormHandler(form_handler.FormHandler):
             result += temp.dot(y)
 
         return result
-
-    def compute_active_sets(self) -> None:
-        """Computes the indices corresponding to active and inactive sets."""
-        self.idx_active_lower.clear()
-        self.idx_active_upper.clear()
-        self.idx_active.clear()
-        self.idx_inactive.clear()
-
-        for j in range(self.control_dim):
-
-            if self.require_control_constraints[j]:
-                self.idx_active_lower.append(
-                    np.nonzero(
-                        (
-                            self.controls[j].vector()[:]
-                            <= self.control_constraints[j][0].vector()[:]
-                        )
-                    )[0]
-                )
-                self.idx_active_upper.append(
-                    np.nonzero(
-                        (
-                            self.controls[j].vector()[:]
-                            >= self.control_constraints[j][1].vector()[:]
-                        )
-                    )[0]
-                )
-                self.idx_inactive.append(
-                    np.nonzero(
-                        np.logical_and(
-                            self.controls[j].vector()[:]
-                            > self.control_constraints[j][0].vector()[:],
-                            self.controls[j].vector()[:]
-                            < self.control_constraints[j][1].vector()[:],
-                        )
-                    )[0]
-                )
-            else:
-                self.idx_active_lower.append([])
-                self.idx_active_upper.append([])
-                self.idx_inactive.append([])
-
-            temp_active = np.concatenate(
-                (self.idx_active_lower[j], self.idx_active_upper[j])
-            )
-            temp_active.sort()
-            self.idx_active.append(temp_active)
-
-    def restrict_to_active_set(
-        self, a: List[fenics.Function], b: List[fenics.Function]
-    ) -> List[fenics.Function]:
-        """Restricts a function to the active set.
-
-        Restricts a control type function ``a`` onto the active set,
-        which is returned via the function ``b``,  i.e., ``b`` is zero on the inactive
-        set.
-
-        Args:
-            a: The first argument, to be projected onto the active set.
-            b: The second argument, which stores the result (is overwritten).
-
-        Returns:
-            The result of the projection (overwrites input b).
-
-        """
-        for j in range(self.control_dim):
-            if self.require_control_constraints[j]:
-                self.temp[j].vector().vec().set(0.0)
-                self.temp[j].vector().apply("")
-                self.temp[j].vector()[self.idx_active[j]] = a[j].vector()[
-                    self.idx_active[j]
-                ]
-                self.temp[j].vector().apply("")
-                b[j].vector().vec().aypx(0.0, self.temp[j].vector().vec())
-                b[j].vector().apply("")
-
-            else:
-                b[j].vector().vec().set(0.0)
-                b[j].vector().apply("")
-
-        return b
-
-    def restrict_to_inactive_set(
-        self, a: List[fenics.Function], b: List[fenics.Function]
-    ) -> List[fenics.Function]:
-        """Restricts a function to the inactive set.
-
-        Restricts a control type function ``a`` onto the inactive set,
-        which is returned via the function ``b``, i.e., ``b`` is zero on the active set.
-
-        Args:
-            a: The control-type function that is to be projected onto the inactive set.
-            b: The storage for the result of the projection (is overwritten).
-
-        Returns:
-            The result of the projection of ``a`` onto the inactive set (overwrites
-            input ``b``).
-
-        """
-        for j in range(self.control_dim):
-            if self.require_control_constraints[j]:
-                self.temp[j].vector().vec().set(0.0)
-                self.temp[j].vector().apply("")
-                self.temp[j].vector()[self.idx_inactive[j]] = a[j].vector()[
-                    self.idx_inactive[j]
-                ]
-                self.temp[j].vector().apply("")
-                b[j].vector().vec().aypx(0.0, self.temp[j].vector().vec())
-                b[j].vector().apply("")
-
-            else:
-                if not b[j].vector().vec().equal(a[j].vector().vec()):
-                    b[j].vector().vec().aypx(0.0, a[j].vector().vec())
-                    b[j].vector().apply("")
-
-        return b
-
-    def project_to_admissible_set(
-        self, a: List[fenics.Function]
-    ) -> List[fenics.Function]:
-        """Project a function to the set of admissible controls.
-
-        Projects a control type function ``a`` onto the set of admissible controls
-        (given by the box constraints).
-
-        Args:
-            a: The function which is to be projected onto the set of admissible
-                controls (is overwritten)
-
-        Returns:
-            The result of the projection (overwrites input ``a``)
-
-        """
-        for j in range(self.control_dim):
-            if self.require_control_constraints[j]:
-                a[j].vector().vec().pointwiseMin(
-                    self.control_constraints[j][1].vector().vec(), a[j].vector().vec()
-                )
-                a[j].vector().apply("")
-                a[j].vector().vec().pointwiseMax(
-                    a[j].vector().vec(), self.control_constraints[j][0].vector().vec()
-                )
-                a[j].vector().apply("")
-
-        return a
 
     def _compute_gradient_equations(self) -> None:
         """Calculates the variational form of the gradient equation."""
