+++ cashocs/_optimization/optimization_variable_abstractions.py
@@ -133,3 +133,53 @@ class OptimizationVariableAbstractions(abc.ABC):
 
         """
         pass
+
+    def compute_active_sets(self) -> None:
+        """Computes the active sets of the problem."""
+        pass
+
+    def restrict_to_inactive_set(
+        self, a: List[fenics.Function], b: List[fenics.Function]
+    ) -> List[fenics.Function]:
+        """Restricts a function to the inactive set.
+
+        Note, that nothing will happen if the type of the optimization problem does not
+        support box constraints.
+
+        Args:
+            a: The function, which shall be restricted to the inactive set
+            b: The output function, which will contain the result and is overridden.
+
+        Returns:
+            The result of the restriction (overrides input b)
+
+        """
+        for j in range(len(self.gradient)):
+            if not b[j].vector().vec().equal(a[j].vector().vec()):
+                b[j].vector().vec().aypx(0.0, a[j].vector().vec())
+                b[j].vector().apply("")
+
+        return b
+
+    # pylint: disable=unused-argument
+    def restrict_to_active_set(
+        self, a: List[fenics.Function], b: List[fenics.Function]
+    ) -> List[fenics.Function]:
+        """Restricts a function to the active set.
+
+        Note, that nothing will happen if the type of the optimization problem does not
+        support box constraints.
+
+        Args:
+            a: The function, which shall be restricted to the active set
+            b: The output function, which will contain the result and is overridden.
+
+        Returns:
+            The result of the restriction (overrides input b)
+
+        """
+        for j in range(len(self.gradient)):
+            b[j].vector().vec().set(0.0)
+            b[j].vector().apply("")
+
+        return b
